+++ src/wrapped_radix_engine_toolkit/high_level/sargon_specific_types/assets_transfers/per_recipient/per_recipient_fungible_transfer.rs
@@ -5,7 +5,7 @@ impl PerRecipientFungibleTransfer {
         resource_address: ResourceAddress,
         amount: impl Into<Decimal192>,
         use_try_deposit_or_abort: bool,
-        divisibility: impl Into<Option<i32>>,
+        divisibility: impl Into<Option<u8>>,
     ) -> Self {
         Self {
             resource_address,
